Extract methods into classes
