+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckerutils/singletracecheck/AssertOrderInsideLoopFirst1.java
@@ -1,0 +1,21 @@
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWord;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.relation.HashTreeRelation;
+
+public class AssertOrderInsideLoopFirst1<L> extends AssertOrder<L> {
+	@Override
+	public List<Set<Integer>> partitionTrace(final NestedWord<L> trace,
+			final List<Object> controlConfigurationSequence) {
+		final HashTreeRelation<Object, Integer> rwt =
+				computeRelationWithTreeSetForTrace(0, trace.length(), controlConfigurationSequence);
+		final Map<Integer, Set<Integer>> depth2Statements =
+				partitionStatementsAccordingDepth(trace, rwt, controlConfigurationSequence);
+		return depth2Statements.keySet().stream().sorted((i1, i2) -> i2.compareTo(i1)).map(depth2Statements::get)
+				.toList();
+	}
+}
