+++ trunk/source/Library-TraceCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/tracecheckerutils/singletracecheck/AssertOrderOutsideLoopFirst1.java
@@ -1,0 +1,27 @@
+package de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWord;
+import de.uni_freiburg.informatik.ultimate.util.datastructures.relation.HashTreeRelation;
+
+public class AssertOrderOutsideLoopFirst1<L> extends AssertOrder<L> {
+	@Override
+	public List<Set<Integer>> partitionTrace(final NestedWord<L> trace,
+			final List<Object> controlConfigurationSequence) {
+		final HashTreeRelation<Object, Integer> rwt =
+				computeRelationWithTreeSetForTrace(0, trace.length(), controlConfigurationSequence);
+		final Map<Integer, Set<Integer>> depth2Statements =
+				partitionStatementsAccordingDepth(trace, rwt, controlConfigurationSequence);
+		// Statements outside of a loop have depth 0.
+		// First, annotate and assert the statements, which doesn't occur within a loop
+		final Set<Integer> stmtsOutsideOfLoop = depth2Statements.get(0);
+		if (stmtsOutsideOfLoop.size() == trace.length()) {
+			return List.of(stmtsOutsideOfLoop);
+		}
+		final Set<Integer> stmtsWithinLoop = getTraceDifference(trace, stmtsOutsideOfLoop);
+		return List.of(stmtsOutsideOfLoop, stmtsWithinLoop);
+	}
+}
